requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via black, virtualenv
asgiref==3.2.10           # via django
attrs==20.1.0             # via black
black==19.10b0            # via -r requirements.in
click==7.1.2              # via black
distlib==0.3.1            # via virtualenv
dj-database-url==0.5.0    # via -r requirements.in
django==3.0.4             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
envparse==0.2.0           # via -r requirements.in
filelock==3.0.12          # via tox, virtualenv
gunicorn==19.9.0          # via -r requirements.in
invoke==1.4.1             # via -r requirements.in
isort==4.3.21             # via -r requirements.in
packaging==20.4           # via tox
pathspec==0.8.0           # via black
pluggy==0.13.1            # via tox
psycopg2-binary==2.8.4    # via -r requirements.in
py==1.9.0                 # via tox
pyparsing==2.4.7          # via packaging
pytz==2020.1              # via django
regex==2020.7.14          # via black
six==1.15.0               # via packaging, tox, virtualenv
sqlparse==0.3.1           # via django
toml==0.10.1              # via black, tox
tox==3.14.5               # via -r requirements.in
typed-ast==1.4.1          # via black
virtualenv==20.0.31       # via tox
